requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-binance>=1.0.19
ccxt>=4.0.0
pyyaml>=6.0
colorama>=0.4.6
python-dotenv>=1.0.0
requests>=2.32.0
yfinance>=0.2.33
